Moved rm to aggregate and updated docs
