+++ bach/docs/source/open-model-hub/models/aggregation/index.rst
@@ -25,6 +25,7 @@ product analytics metrics. Always return aggregated data in some form from the
     unique_sessions
     session_duration
     frequency
+    retention_matrix
     top_product_features
     top_product_features_before_conversion
