+++ modelhub/modelhub/aggregate.py
@@ -6,7 +6,7 @@ from typing import List, Union
 import bach
 from bach.series import Series
 from sql_models.constants import NotSet, not_set
-from typing import List, Union, TYPE_CHECKING
+from typing import cast, List, Union, TYPE_CHECKING
 
 
 if TYPE_CHECKING:
@@ -280,3 +280,190 @@ class Aggregate:
         converted_users_features = converted_users_features.dropna().set_index(_index)
 
         return converted_users_features.sort_values('unique_users', ascending=False)
+
+    def retention_matrix(self,
+                         data: bach.DataFrame,
+                         time_period: str = 'monthly',
+                         event_type: str = None,
+                         start_date: str = None,
+                         end_date: str = None,
+                         percentage=False,
+                         display=True) -> bach.DataFrame:
+
+        """
+        Finds the number of users in a given cohort who are active at a given time
+        period, where time is computed with respect to the beginning of each cohort.
+        The "active user" is the user who made an action that we are interested in
+        that time period.
+        Users are divided into mutually exclusive cohorts, which are then
+        tracked over time. In our case users are assigned to a cohort based on
+        when they made their first action that we are interested in.
+
+        Returns the retention matrix dataframe, it represents users retained across cohorts:
+
+        - index value represents the cohort
+        - columns represent the number of given date period since the current cohort
+        - values represent number (or percentage) of unique active users of a given cohort
+
+        One can calculate the retention matrix for a given time range, for that
+        one can specify start_date a/o end_date.
+        N.B. the users' activity starts to be traced from the first date the user is seen in the `data`.
+
+
+        :param data: :py:class:`bach.DataFrame` to apply the method on.
+        :param time_period: can be 'daily', 'weekly', 'monthly' or 'yearly'.
+        :param event_type: the event/action that we are interested in.
+            Must be a valid event_type (either parent or child).
+            if None we take all the events generated by the user.
+        :param start_date: start date of the retention matrix, e.g. '2022-04-01'
+            if None take all the data.
+        :param end_date: end date of the retention matrix, e.g. '2022-05-01'
+            if None take all the data.
+        :param percentage: if True calculate percentage with respect to the number of a users
+            in the cohort, otherwise it leaves the absolute values.
+        :param display: if display==True visualize the retention matrix as a heat map
+
+        :returns: retention matrix bach DataFrame.
+        """
+
+        available_formats = {'daily', 'weekly', 'monthly', 'yearly'}
+
+        if time_period not in available_formats:
+            raise ValueError(f'{time_period} time_period is not available.')
+
+        from datetime import datetime
+
+        _start_date = None
+        if start_date is not None:
+            try:
+                _start_date = datetime.strptime(start_date, '%Y-%m-%d')
+            except Exception as e:
+                print('Please provide correct start_date formatted as "YYYY-MM-DD".')
+                raise e
+
+        _end_date = None
+        if end_date is not None:
+            try:
+                _end_date = datetime.strptime(end_date, '%Y-%m-%d')
+            except Exception as e:
+                print('Please provide correct end_date formatted as "YYYY-MM-DD".')
+                raise e
+
+        self._mh._check_data_is_objectiv_data(data)
+
+        data = data.copy()
+
+        # filtering data with the event that we are interested in
+        if event_type is not None:
+            data = data[data['event_type'] == event_type]
+
+        # for retention matrix calculation we need only event date and user_id
+        columns = ['user_id', 'moment']
+        data = data[columns]
+
+        # get the first cohort
+        cohorts = data.groupby('user_id')['moment'].min().reset_index()
+        cohorts = cohorts.rename(columns={'moment': 'first_cohort_ts'})
+
+        # add first cohort to our data DataFrame
+        data = data.merge(cohorts, on='user_id')
+
+        # help mypy
+        from bach import SeriesTimestamp, SeriesTimedelta
+        moment = cast(SeriesTimestamp, data['moment'])
+        first_cohort_ts = cast(SeriesTimestamp, data['first_cohort_ts'])
+
+        # calculate cohort distance
+        if time_period == 'yearly':
+            data['cohort'] = moment.dt.strftime('%Y').astype(dtype=int)
+            data['first_cohort_yearly'] = first_cohort_ts.dt.strftime('%Y').astype(dtype=int)
+            data['cohort_distance'] = data['cohort'] - data['first_cohort_yearly']
+
+            data['first_cohort'] = first_cohort_ts.dt.strftime('%Y')
+        elif time_period == 'monthly':
+            data['cohort_year'] = moment.dt.strftime('%Y').astype(dtype=int)
+            data['first_cohort_year'] = first_cohort_ts.dt.strftime('%Y').astype(dtype=int)
+            data['cohort_year_diff'] = data['cohort_year'] - data['first_cohort_year']
+
+            data['cohort_month'] = moment.dt.strftime('%m').astype(dtype=int)
+            data['first_cohort_month'] = first_cohort_ts.dt.strftime('%m').astype(dtype=int)
+            data['cohort_month_diff'] = data['cohort_month'] - data['first_cohort_month']
+
+            n_months = 12
+            data['cohort_distance'] = data['cohort_year_diff'] * n_months + data['cohort_month_diff']
+
+            data['first_cohort'] = first_cohort_ts.dt.strftime('%Y-%m')
+        elif time_period == 'weekly':
+            n_days = 7.0
+
+            data['cohort'] = moment.dt.date_trunc('week').astype('timestamp')
+            data['first_cohort_weekly'] = first_cohort_ts.dt.date_trunc('week')
+            cohort_distance = cast(SeriesTimedelta, data['cohort'] - data['first_cohort_weekly'])
+            data['cohort_distance'] = cohort_distance.dt.days
+            data['cohort_distance'] = data['cohort_distance'] / n_days
+
+            pretty_name = cast(SeriesTimestamp, data['first_cohort_weekly']).dt.strftime('%Y-%m-%d')
+            data['first_cohort'] = pretty_name
+        else:
+            # daily
+            data['cohort_distance'] = data['moment'] - data['first_cohort_ts']
+            data['cohort_distance'] = cast(SeriesTimedelta, data['cohort_distance']).dt.days
+            data['first_cohort'] = first_cohort_ts.dt.strftime('%Y-%m-%d')
+
+        # applying start date filter
+        if _start_date is not None:
+            data = data[data['first_cohort_ts'] >= _start_date]
+
+        # applying end date filter
+        if _end_date is not None:
+            data = data[data['moment'] < _end_date]
+
+        # in BigQuery columns cannot start with numbers
+        data['cohort_distance_prefix'] = '_'
+        data['cohort_distance'] = data['cohort_distance'].astype(dtype=str)
+        data['cohort_distance'] = data['cohort_distance_prefix'] + data['cohort_distance']
+
+        retention_matrix = data.groupby(['first_cohort',
+                                         'cohort_distance']).agg({'user_id': 'nunique'}).unstack(
+            level='cohort_distance')
+
+        # renaming columns, removing string attached after unstacking
+        column_name_map = {col: col.replace('__user_id_nunique', '').replace('.0', '')
+                           for col in retention_matrix.data_columns}
+        retention_matrix = retention_matrix.rename(columns=column_name_map)
+
+        # 'sort' with column names (numerical sorting, even though the columns are strings)
+        columns = [f'_{j}' for j in sorted([int(i.replace('_', ''))
+                                            for i in retention_matrix.data_columns])]
+        retention_matrix = retention_matrix[columns]
+        # for BigQuery we need sorting
+        retention_matrix = retention_matrix.sort_index()
+
+        if percentage and len(columns):
+            first_column = retention_matrix[columns[0]]
+            for col in columns:
+                retention_matrix[col] = (retention_matrix[col] / first_column) * 100
+
+        if display and len(columns):
+            import matplotlib.pyplot as plt
+            import seaborn as sns
+            fig, ax = plt.subplots(figsize=(20, 8))
+            sns.heatmap(retention_matrix.to_pandas(), annot=True, square=True, ax=ax,
+                        linewidths=.5, cmap=sns.cubehelix_palette(rot=-.4), fmt='.1f')
+            if percentage:
+                [t.set_text(f'{t.get_text()}%') for t in ax.texts]
+
+            plt.title('Cohort Analysis')
+
+            nice_name = {
+                'daily': 'Days',
+                'weekly': 'Weeks',
+                'monthly': 'Months',
+                'yearly': 'Years'
+            }
+
+            plt.xlabel(f'{nice_name[time_period]} After First Event')
+            plt.ylabel('First Event Cohort')
+            plt.show()
+
+        return retention_matrix
